+++ harvester/__init__.py
@@ -4,7 +4,6 @@ import os
 from dotenv import load_dotenv
 from sqlalchemy import create_engine
 from sqlalchemy.orm import scoped_session, sessionmaker
-from sqlalchemy.pool import NullPool
 
 from config.logger_config import LOGGING_CONFIG
 from database.interface import HarvesterDBInterface
@@ -16,7 +15,7 @@ logging.config.dictConfig(LOGGING_CONFIG)
 DATABASE_URI = os.getenv("DATABASE_URI")
 
 # create a scopedsession for our harvest runner
-engine = create_engine(DATABASE_URI, poolclass=NullPool)
+engine = create_engine(DATABASE_URI)
 session_factory = sessionmaker(bind=engine, autoflush=True)
 session = scoped_session(session_factory)
