+++ tests/integration/harvest/test_ckan_load.py
@@ -14,7 +14,7 @@ from harvester.utils.general_utils import dataset_to_hash, sort_dataset
 
 
 class TestCKANLoad:
-    @patch("harvester.harvest.ckan")
+    @patch("harvester.harvest.ckan_sync_tool.ckan")
     def test_sync(
         self,
         CKANMock,
@@ -55,7 +55,7 @@ class TestCKANLoad:
         results = {
             "action": {"create": 0, "update": 0, "delete": 0, None: 0},
             "status": {"success": 0, "error": 0, None: 0},
-            "validity": {True: 0, False: 0},
+            "validity": {True: 0, None: 0},
         }
         for key, group in groupby(
             harvest_source.records, lambda x: x.action if x.status != "error" else None
@@ -65,21 +65,32 @@ class TestCKANLoad:
         for key, group in groupby(harvest_source.records, lambda x: x.status):
             results["status"][key] = sum(1 for _ in group)
 
+        for key, group in groupby(harvest_source.records, lambda x: x.valid):
+            results["validity"][key] = sum(1 for _ in group)
+
         harvest_reporter = harvest_source.reporter.report()
         assert results["action"]["create"] == 6
         assert harvest_reporter["records_added"] == 6
 
-        assert results["action"]["update"] == 1
-        assert harvest_reporter["records_updated"] == 1
+        assert results["action"]["update"] == harvest_reporter["records_updated"] == 1
 
-        assert results["action"]["delete"] == 1
-        assert harvest_reporter["records_deleted"] == 1
+        assert results["action"]["delete"] == harvest_reporter["records_deleted"] == 1
 
-        assert results["status"]["error"] == 0
-        assert harvest_reporter["records_errored"] == 0
+        assert results["status"]["error"] == harvest_reporter["records_errored"] == 0
 
-        assert results["status"]["success"] == 8
-        assert len(harvest_source.records) - harvest_reporter["records_errored"] == 8
+        assert (
+            results["status"]["success"]
+            == len(harvest_source.records) - harvest_reporter["records_errored"]
+            == 8
+        )
+
+        assert results["validity"][True] == harvest_reporter["records_validated"] == 7
+
+        assert (
+            results["validity"][None]
+            == len(harvest_source.records) - harvest_reporter["records_validated"]
+            == 1
+        )
 
     def test_ckanify_dcatus(
         self,
